Create create checklist page/view
The ability to create checklists on the create checklist page has been added. You can access the create checklist page via the home page. The
fields in the create checklist page do not have an effect on the checklist posted. Once a checklist is posted an error will appear and you can go
back to the homepage, refresh, and it will be there.